pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hr-growth-assessment"
version = "0.1.0"
description = "Config-driven HR growth-areas assessment tool for solo consultants"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "jsonschema>=4.20",
  "PyYAML>=6.0",
  "openai>=1.40",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "readability-lxml>=0.8.1",
  "pypdf>=4.2",
  "python-docx>=1.1",
  "jinja2>=3.1",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2"
]

[tool.setuptools]
packages = ["app", "app.ingest", "app.llm", "app.logic", "app.report", "app.web"]

[tool.setuptools.package-data]
"app.web" = ["static/*.html", "static/*.css", "static/*.js"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
